requirements.txt
Python==3.7.3
pip==19.1.1
django-admin==2.2.1
djangorestframework== 3.9.3
pytest==4.4.2
pytest-metadata==1.8.0
pytest-html==1.20.0
django-simple==log-0.3.4
psycopg2-binary==2.8.2